Combine the Domain Model with the Providers::Base
`Providers::Base::Domain` and `Models::Domain` where significantly
coupled. To reduce this, they have been collapsed down into a single
`Providers::Base::Domain` class.

Simple inheritance can then be used to distinguish between the different
providers. The `Providers.select` method can be used to dynamically
select the correct provider.

As inheritance is being used to distinguish the providers, the `provider`
attribute needed to be removed. Instead it is hard coded in the
inherited class.

NOTE: This commit has broken the machine commands. However the `Machine`
model needs refactoring in the same way as `Domain`
