+++ spec/factories.rb
@@ -2,11 +2,10 @@
 # frozen_string_literal: true
 
 FactoryBot.define do
-  models = Cloudware::Models
+  aws = Cloudware::Providers::AWS
 
-  factory :domain, class: models::Domain do
+  factory :aws_domain, class: aws::Domain do
     name 'Test-Domain-Name-1'
-    provider 'aws'
     region 'eu-west-1'
     networkcidr '10.0.0.0/16'
     prisubnetcidr '10.0.1.0/24'
